+++ src/main/java/ru/vyarus/guice/persist/orient/db/transaction/template/TxTemplate.java
@@ -73,7 +73,7 @@ public class TxTemplate {
     }
 
     private void throwRuntime(final Throwable th) {
-        Throwables.propagateIfPossible(th);
+        Throwables.throwIfUnchecked(th);
         throw new TemplateTransactionException("Transaction template execution failed", th);
     }